Refactor: duplication
